backend/requirements.txt
alembic==1.16.2
blinker==1.9.0
click==8.2.1
dotenv==0.9.9
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.0.0
psycopg==3.1.18
redis==5.0.1
celery==5.3.6
sentence-transformers==2.2.2
faiss-cpu==1.7.4
requests==2.31.0
pdfplumber==0.10.3
python-magic==0.4.27
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
